+++ drizzle/20260305233323_optOut/migration.sql
@@ -1,0 +1,5 @@
+CREATE TABLE `optOut` (
+	`user_id` text PRIMARY KEY,
+	`createdAt` text NOT NULL,
+	`updatedAt` text NOT NULL
+);