+++ changelog.md
@@ -1,3 +1,14 @@
+## v8.3.0 - Mar 07, 2026
+
+### Enhancements
+
+- Update wording on homepage to be login aware
+- Fix bug in user stats that was showing the guild ID instead of the name
+- Added option for GDPR data deletions in the user settings page
+- Added an option to opt out of interactions via the user settings
+- Added a very basic privacy policy
+- Added a very basic terms of service
+
 ## v8.2.5 - Feb 22, 2026
 
 ### Enhancements